+++ app/src/Controller/UserController.php
@@ -295,6 +295,7 @@ class UserController extends AbstractController
     }
 
     #[Route(path: '/email/{id}/delete', name: 'user_email_delete', methods: 'POST')]
+    #[IsGranted('DOMAIN_ACCESS', subject: 'user')]
     public function deleteEmail(Request $request, User $user): Response
     {
         if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->getString('_token'))) {
@@ -318,7 +319,7 @@ class UserController extends AbstractController
 
         foreach ($request->request->all('id') as $id) {
             $user = $this->em->getRepository(User::class)->find($id);
-            if ($user) {
+            if ($user && $this->isGranted('DOMAIN_ACCESS', $user)) {
                 $this->em->remove($user);
             }
         }
@@ -496,6 +497,7 @@ class UserController extends AbstractController
 
 
     #[Route(path: '/email/{id}/edit', name: 'user_email_edit', methods: 'GET|POST')]
+    #[IsGranted('DOMAIN_ACCESS', subject: 'user')]
     public function editUserEmail(
         User $user,
         Request $request,
@@ -689,7 +691,7 @@ class UserController extends AbstractController
     {
         /** @var User $user */
         $user = $this->getUser();
-        if (!in_array('ROLE_SUPER_ADMIN', $user->getRoles())) {
+        if (!$this->isGranted('ROLE_SUPER_ADMIN')) {
             $allowedomains = $user->getDomains()->toArray();
         } else {
             $allowedomains = $this->em
@@ -730,7 +732,7 @@ class UserController extends AbstractController
 
         /** @var User $user */
         $user = $this->getUser();
-        if (!in_array('ROLE_SUPER_ADMIN', $user->getRoles())) {
+        if (!$this->isGranted('ROLE_SUPER_ADMIN')) {
             $allowedomains = $user->getDomains()->toArray();
         }
